# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    Proof of Personhood Protocol API.

    All `/api/*` endpoints (except `/healthz` and `/readyz`) require a project
    API key sent as `Authorization: Bearer pk_test_…` or
    `Authorization: Bearer pk_live_…`.

    Mutating endpoints accept an optional `Idempotency-Key` header. Sending the
    same key with the same body within 24 hours returns the original response
    byte-for-byte.

    Errors always use the `ErrorEnvelope` shape, with a stable
    machine-readable `error.code` and the `request_id` for log correlation.
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: protocol
    description: Proof of Personhood Protocol operations
security:
  - bearerAuth: []
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check (with DB ping)
      description: Returns server health status. Returns 503 if a critical dependency (Postgres) is not reachable.
      security: []
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "503":
          description: One or more critical dependencies are unhealthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /readyz:
    get:
      operationId: readinessCheck
      tags: [health]
      summary: Readiness probe
      description: |
        Returns 200 with `status: "ready"` once the server can reach Postgres
        and the expected schema (core protocol tables + hardening tables) is
        present. Returns 503 with `status: "not_ready"` otherwise. Suitable
        for orchestrator readiness gates.
      security: []
      responses:
        "200":
          description: Server ready to accept traffic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"
        "503":
          description: Server not ready (DB unreachable or schema incomplete)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"

  /inquiries:
    post:
      operationId: createInquiry
      tags: [protocol]
      summary: Start a hosted liveness inquiry
      description: |
        Creates a hosted identity-verification inquiry with the configured
        vendor (Persona by default; a "mock" vendor activates automatically
        when Persona credentials are absent so the demo flow keeps working).
        Returns a `hostedUrl` that you redirect the end user to.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInquiryRequest"
      responses:
        "200":
          description: Inquiry created
          headers:
            X-Request-ID: { $ref: "#/components/headers/XRequestId" }
            X-RateLimit-Limit: { $ref: "#/components/headers/XRateLimitLimit" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/XRateLimitRemaining" }
            X-RateLimit-Reset: { $ref: "#/components/headers/XRateLimitReset" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InquirySession"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502":
          description: Identity vendor is unreachable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /inquiries/{inquiryId}:
    get:
      operationId: getInquiry
      tags: [protocol]
      summary: Poll inquiry status
      description: Returns the current status of an inquiry. Use this to wait until the user has completed the hosted flow before calling /register.
      parameters:
        - name: inquiryId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Inquiry status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InquiryStatus"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Inquiry not found for this project.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "429": { $ref: "#/components/responses/RateLimited" }

  /register:
    post:
      operationId: registerCommitment
      tags: [protocol]
      summary: Register a verified human commitment
      description: |
        Mints a Proof-of-Personhood commitment from a completed inquiry.
        The server derives a per-app nullifier (HMAC-SHA256) and signs an
        RS256 human badge JWT. Verifiers can validate the badge offline
        using the JWKS at `/.well-known/jwks.json`.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Commitment registered successfully
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "408": { $ref: "#/components/responses/RequestTimeout" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }

  /verify:
    post:
      operationId: verifyProof
      tags: [protocol]
      summary: Verify a human-badge JWT
      description: |
        Verifies an RS256-signed human badge JWT. Checks the signature
        against the JWKS, the issuer, the audience (your project id), the
        expiration, and the app_context claim. Cross-checks the embedded
        nullifier against the on-chain-style commitment registry.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Verification result
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "408": { $ref: "#/components/responses/RequestTimeout" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }

  /stats:
    get:
      operationId: getProtocolStats
      tags: [protocol]
      summary: Get global protocol statistics
      description: Returns aggregate statistics about the protocol.
      responses:
        "200":
          description: Protocol statistics
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtocolStats"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }

  /nullifier/{hash}:
    get:
      operationId: checkNullifier
      tags: [protocol]
      summary: Check if a nullifier has been used
      description: Returns whether a given nullifier hash has already been consumed.
      parameters:
        - name: hash
          in: path
          required: true
          description: The nullifier hash to check
          schema:
            type: string
      responses:
        "200":
          description: Nullifier status
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NullifierStatus"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key (pk_test_… / pk_live_…)

  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Opaque idempotency key (≤200 chars). Repeating the same key + body within 24h returns the original response.
      schema:
        type: string
        maxLength: 200

  headers:
    XRequestId:
      description: Unique id for this request, useful for support / log correlation.
      schema: { type: string }
    XRateLimitLimit:
      description: Maximum requests for this bucket per minute.
      schema: { type: integer }
    XRateLimitRemaining:
      description: Tokens remaining in the current bucket window.
      schema: { type: integer }
    XRateLimitReset:
      description: Seconds until the bucket is fully refilled.
      schema: { type: integer }
    RetryAfter:
      description: Seconds to wait before retrying.
      schema: { type: integer }

  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Origin not allowed for this live API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    RequestTimeout:
      description: Server-side request timeout.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: Conflicting state (duplicate nullifier or replayed Idempotency-Key with a different body).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    PayloadTooLarge:
      description: Request body exceeded the 32 KB limit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ValidationError:
      description: Request body did not pass schema validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    RateLimited:
      description: Per-project rate limit exceeded.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        X-RateLimit-Limit:
          $ref: "#/components/headers/XRateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/XRateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/XRateLimitReset"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ReadinessStatus:
      type: object
      properties:
        status:
          type: string
          enum: [ready, not_ready]
        checks:
          type: object
          description: |
            Per-dependency readiness signal. `db` is the Postgres ping;
            `schema` confirms the expected tables exist (the project does
            not use a separate migration runner — schema is bootstrapped on
            startup, and this check verifies that bootstrap completed).
      required:
        - status

    CreateInquiryRequest:
      type: object
      properties:
        referenceId:
          type: string
          description: Optional opaque id you can use to correlate this inquiry back to your own user record.
          maxLength: 128
        redirectUri:
          type: string
          format: uri
          description: Where the hosted flow should send the user after completion.
          maxLength: 512

    InquirySession:
      type: object
      properties:
        inquiryId: { type: string }
        hostedUrl:
          type: string
          description: Open this URL in a browser to start the hosted liveness check.
        vendor:
          type: string
          enum: [mock, persona]
        status:
          type: string
          enum: [pending]
      required: [inquiryId, hostedUrl, vendor, status]

    InquiryStatus:
      type: object
      properties:
        inquiryId: { type: string }
        status:
          type: string
          enum: [pending, completed, approved, declined, expired, failed]
        decision:
          type: string
          nullable: true
          enum: [approved, declined, null]
      required: [inquiryId, status]

    RegisterRequest:
      type: object
      properties:
        inquiryId:
          type: string
          description: Id returned by POST /inquiries after the user completes the hosted flow.
        appContext:
          type: string
          description: Application context. Same (subject, appContext) ⇒ same nullifier; different appContexts are uncorrelated.
          maxLength: 128
      required: [inquiryId, appContext]

    RegisterResponse:
      type: object
      properties:
        commitmentHash:
          type: string
          description: HMAC-SHA256 commitment id. Cannot be reversed to recover the subject.
        nullifier:
          type: string
          description: HMAC-SHA256 nullifier scoped to (subject, appContext).
        humanBadge:
          type: string
          description: RS256-signed JWT. Verify with JWKS at /.well-known/jwks.json.
        registeredAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          description: When the issued human badge expires (24h).
      required: [commitmentHash, nullifier, humanBadge, registeredAt, expiresAt]

    VerifyRequest:
      type: object
      properties:
        humanBadge:
          type: string
          description: The JWT returned by /register.
        appContext:
          type: string
          description: Must match the app_context claim in the badge.
      required: [humanBadge, appContext]

    VerifyResponse:
      type: object
      properties:
        verified: { type: boolean }
        nullifier:
          type: string
          description: Echoed from the badge claims; absent on failure.
        commitmentHash:
          type: string
          description: Echoed from the badge subject; absent on failure.
        verifiedAt:
          type: string
          format: date-time
        message: { type: string }
      required: [verified, verifiedAt, message]

    ProtocolStats:
      type: object
      properties:
        totalCommitments:
          type: integer
        totalVerifications:
          type: integer
        totalFailedVerifications:
          type: integer
        uptimeSeconds:
          type: number
        activeNullifiers:
          type: integer
      required:
        - totalCommitments
        - totalVerifications
        - totalFailedVerifications
        - uptimeSeconds
        - activeNullifiers

    NullifierStatus:
      type: object
      properties:
        hash:
          type: string
        used:
          type: boolean
        registeredAt:
          type: string
          format: date-time
      required:
        - hash
        - used

    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
              enum:
                - missing_authorization
                - invalid_api_key
                - revoked_api_key
                - forbidden_origin
                - rate_limited
                - idempotency_conflict
                - idempotency_in_progress
                - payload_too_large
                - request_timeout
                - validation_error
                - not_found
                - conflict
                - internal_error
                - service_unavailable
                - inquiry_not_found
                - inquiry_not_approved
                - inquiry_consumed
                - invalid_badge
                - expired_badge
                - vendor_unavailable
                - webhook_signature_invalid
            message:
              type: string
              description: Human-readable explanation suitable for surfacing to developers.
            request_id:
              type: string
              description: The X-Request-ID for this response. Quote this when contacting support.
            details:
              type: string
              description: Optional extra context (e.g. Zod validation message).
          required:
            - code
            - message
            - request_id
      required:
        - error